dop done

+++ src/main/java/is/lab1/repository/CityRepository.java
@@ -59,8 +59,8 @@ public interface CityRepository extends JpaRepository<City, Integer> {
     Page<City> findAll(Pageable pageable);
     
     @Query("SELECT c FROM City c WHERE c.coordinates.x = :x AND c.coordinates.y = :y")
-    List<City> findByCoordinates(@Param("x") Double x, @Param("y") Float y);
+    List<City> findByCoordinates(@Param("x") Integer x, @Param("y") Integer y);
     
     @Query("SELECT c FROM City c WHERE c.coordinates.x = :x AND c.coordinates.y = :y AND c.id != :cityId")
-    List<City> findByCoordinatesExcludingCity(@Param("x") Double x, @Param("y") Float y, @Param("cityId") Integer cityId);
+    List<City> findByCoordinatesExcludingCity(@Param("x") Integer x, @Param("y") Integer y, @Param("cityId") Integer cityId);
 }
